Add tests, fix linter errors
