--- services/observer/notifier/notifier_test.go
@@ -1,1 +1,0 @@
-package notifier
